+++ src/parsing/syntax.ml
@@ -10,6 +10,23 @@ open Core
 (** Representation of a (located) identifier. *)
 type p_ident = strloc
 
+(** [notin id idopts] checks that [id] does not occur in [idopts]. *)
+let check_notin : string -> p_ident option list -> unit = fun id ->
+  let rec notin = function
+    | [] -> ()
+    | None :: idopts -> notin idopts
+    | Some {elt=id';pos} :: idopts ->
+        if id' = id then Error.fatal pos "%s already used." id
+        else notin idopts
+  in notin
+
+(** [are_distinct idopts] checks that the elements of [idopts] of the form
+   [Some _] are pairwise distinct. *)
+let rec check_distinct : p_ident option list -> unit = function
+  | [] -> ()
+  | None :: idopts -> check_distinct idopts
+  | Some {elt=id;_} :: idopts -> check_notin id idopts; check_distinct idopts
+
 (** Identifier of a metavariable. *)
 type meta_ident = Name of string | Numb of int
 type p_meta_ident = meta_ident loc